:recycle: move `DataFrame` formatting to \data_formatting.py
